Move prevent_request_warning decorator to core/test_utils.py and apply more widely to tidy up test output
