+++ etna/ciim/utils.py
@@ -1,4 +1,3 @@
-import logging
 import re
 
 from typing import Any, Dict, Optional
@@ -250,24 +249,3 @@ def format_link(link_html: str) -> Dict[str, str]:
     """
     document = pq(link_html)
     return {"iaid": document.attr("href"), "text": document.text()}
-
-
-def prevent_request_warnings(original_function):
-    """
-    If we need to test for 503s this decorator can prevent the
-    request class from throwing warnings.
-    """
-
-    def new_function(*args, **kwargs):
-        # raise logging level to CRITICAL
-        logger = logging.getLogger("django.request")
-        previous_logging_level = logger.getEffectiveLevel()
-        logger.setLevel(logging.CRITICAL)
-
-        # trigger original function that would throw warning
-        original_function(*args, **kwargs)
-
-        # lower logging level back to previous
-        logger.setLevel(previous_logging_level)
-
-    return new_function
